Reorganize unit tests for clarity
No logical changes intended
