+++ microcosm_postgres/tests/fixtures/__init__.py
@@ -1,0 +1,2 @@
+from microcosm_postgres.tests.fixtures.company import Company, CompanyType  # noqa: F401
+from microcosm_postgres.tests.fixtures.employee import Employee  # noqa: F401
