+++ microcosm_postgres/tests/fixtures/employee.py
@@ -1,33 +1,9 @@
-"""
-Example models and store usage.
-
-"""
-from enum import Enum, unique
-
 from sqlalchemy import Column, ForeignKey, String
 from sqlalchemy_utils import UUIDType
 
 from microcosm.api import binding
 from microcosm_postgres.models import EntityMixin, Model
 from microcosm_postgres.store import Store
-from microcosm_postgres.types import EnumType
-
-
-@unique
-class CompanyType(Enum):
-    private = "private"
-    public = "public"
-
-
-class Company(EntityMixin, Model):
-    """
-    A company has a unique name.
-
-    """
-    __tablename__ = "company"
-
-    name = Column(String(255), unique=True)
-    type = Column(EnumType(CompanyType))
 
 
 class Employee(EntityMixin, Model):
@@ -47,11 +23,12 @@ class Employee(EntityMixin, Model):
         yield (self.company_id, self.id)
 
 
-class CompanyStore(Store):
-    pass
-
-
+@binding("employee_store")
 class EmployeeStore(Store):
+
+    def __init__(self, graph):
+        super().__init__(graph, Employee)
+
     def search_by_company(self, company_id):
         return self.search(Employee.company_id == company_id)
 
@@ -66,13 +43,3 @@ class EmployeeStore(Store):
         if first is not None:
             query = query.filter(Employee.first == first)
         return super(EmployeeStore, self)._filter(query, **kwargs)
-
-
-@binding("company_store")
-def configure_company_store(graph):
-    return CompanyStore(graph, Company, [Company.name])
-
-
-@binding("employee_store")
-def configure_employee_store(graph):
-    return EmployeeStore(graph, Employee)
