+++ microcosm_postgres/tests/test_cloning.py
@@ -7,7 +7,7 @@ from microcosm.api import create_object_graph
 
 from microcosm_postgres.cloning import clone
 from microcosm_postgres.context import SessionContext, transaction
-from microcosm_postgres.tests.example import Company, CompanyType
+from microcosm_postgres.tests.fixtures import Company, CompanyType
 
 
 class TestCloning:
